+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-- Add support for PEP 742, adding `typing_extensions.TypeNarrower`. Patch
+- Add support for PEP 742, adding `typing_extensions.TypeIs`. Patch
   by Jelle Zijlstra.
 - Speedup `issubclass()` checks against simple runtime-checkable protocols by
   around 6% (backporting https://github.com/python/cpython/pull/112717, by Alex